+++ project/stampee/src/main/java/Main.java
@@ -12,36 +12,36 @@ import javafx.stage.Stage;
 
 
 public class Main extends Application {
-	@Override
-	public void start(Stage primaryStage) {
-		try {
-			// FXML 파일 경로를 설정합니다.
-			FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/CouponPage.fxml"));
-
-			// FXML 파일을 로드하고 루트 레이아웃을 가져옵니다.
-			Parent root = loader.load();
-
-			// Scene을 설정하고 Stage에 추가합니다.
-			Scene scene = new Scene(root);
-			primaryStage.setScene(scene);
-			primaryStage.setTitle("Coupon Application");
-			primaryStage.show();
-		} catch(Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	// public void start(Stage primaryStage) throws Exception {
-	// 	FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/stamp.fxml"));
-	// 	Parent root = loader.load();
-	// 	StampController controller = loader.getController();
-	// 	controller.initData(39L, 1L); // 예시로 memberId 1을 사용
-	// 	primaryStage.setTitle("Stamp Viewer");
-	// 	primaryStage.setScene(new Scene(root));
+	// @Override
+	// public void start(Stage primaryStage) {
+	// 	try {
+	// 		// FXML 파일 경로를 설정합니다.
+	// 		FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/CouponPage.fxml"));
 	//
-	// 			primaryStage.setTitle("Coupon Application");
-	// 			primaryStage.show();
+	// 		// FXML 파일을 로드하고 루트 레이아웃을 가져옵니다.
+	// 		Parent root = loader.load();
+	//
+	// 		// Scene을 설정하고 Stage에 추가합니다.
+	// 		Scene scene = new Scene(root);
+	// 		primaryStage.setScene(scene);
+	// 		primaryStage.setTitle("Coupon Application");
+	// 		primaryStage.show();
+	// 	} catch(Exception e) {
+	// 		e.printStackTrace();
+	// 	}
 	// }
+
+	public void start(Stage primaryStage) throws Exception {
+		FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/stamp.fxml"));
+		Parent root = loader.load();
+		StampController controller = loader.getController();
+		controller.initData(39L, 1L); // 예시로 memberId 1을 사용
+		primaryStage.setTitle("Stamp Viewer");
+		primaryStage.setScene(new Scene(root));
+
+				primaryStage.setTitle("Coupon Application");
+				primaryStage.show();
+	}
 
 	// @Override
 	// public void start(Stage primaryStage) throws Exception{
