+++ project/stampee/src/main/java/controller/CouponController.java
@@ -1,6 +1,4 @@
 package controller;
-
-import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
 import java.util.List;
@@ -12,15 +10,10 @@ import domain.Member;
 import dto.response.MyCouponDto;
 import formatter.PhoneNumberFormatter;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
 import repository.CouponRepository;
 import repository.MemberRepository;
 import repository.StampRepository;
@@ -35,8 +28,8 @@ public class CouponController implements Initializable {
 	private final CouponService couponService;
 	private final StampService stampService;
 	private final UserService userService;
-	private final CouponView couponView = new CouponView();
-	private final PopupView popupView = new PopupView();
+	private final CouponView couponView;
+	private final PopupView popupView;
 
 	@FXML private VBox cafeListContainer;
 
@@ -45,6 +38,9 @@ public class CouponController implements Initializable {
 		StampRepository stampRepository = new StampRepository();
 		MemberRepository memberRepository = new MemberRepository();
 		MailService mailService = new MailService();
+
+		couponView = new CouponView();
+		popupView = new PopupView();
 
 		couponService = new CouponService(couponRepository, mailService);
 		stampService = new StampService(stampRepository, memberRepository, mailService);
@@ -69,29 +65,11 @@ public class CouponController implements Initializable {
 	public void handleShareButtonPress(long memberId, long cafeId) {
 		Member fromMember = userService.findMemberById(memberId);
 		try {
-			String toPhoneNumber = PhoneNumberFormatter.formatPhoneNumber(showNumberPadPopup());
+			String toPhoneNumber = PhoneNumberFormatter.formatPhoneNumber(couponView.showNumberPadPopup());
 			stampService.shareStamp(fromMember, cafeId, toPhoneNumber, 1);
 			popupView.showSuccessPopup("스탬프 공유 성공");
 		} catch (MessagingException | SQLException |  IllegalArgumentException e) {
 			popupView.showFailPopup(e.getMessage());
-		}
-	}
-
-	public String showNumberPadPopup() {
-		try {
-			FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/numberPad.fxml"));
-			Parent root = loader.load();
-			NumberPadController controller = loader.getController();
-
-			Stage stage = new Stage();
-			stage.initModality(Modality.APPLICATION_MODAL);
-			stage.setTitle("전화번호 입력");
-			stage.setScene(new Scene(root));
-			stage.showAndWait();
-
-			return controller.getInputField().getText();
-		} catch (Exception e) {
-			return null;
 		}
 	}
