--- project/stampee/src/main/java/controller/ShareCompleteController.java
@@ -1,14 +1,0 @@
-package controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.Node;
-import javafx.stage.Stage;
-
-public class ShareCompleteController {
-	@FXML
-	private void handleClose(ActionEvent event) {
-		Stage stage = (Stage) ((Node)event.getSource()).getScene().getWindow();
-		stage.close();
-	}
-}
