+++ project/stampee/src/main/java/util/PasswordUtil.java
@@ -4,7 +4,6 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.Base64;
-import java.util.Scanner;
 
 public class PasswordUtil {
