+++ project/stampee/src/main/resources/fxml/shareComplete.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.layout.StackPane?>
 
-<StackPane xmlns:fx="http://javafx.com/fxml" fx:controller="controller.ShareCompleteController"
+<StackPane xmlns:fx="http://javafx.com/fxml" fx:controller="controller.NumberPadController"
            styleClass="root" prefWidth="250" prefHeight="200" stylesheets="/css/shareComplete.css">
     <VBox alignment="CENTER" spacing="20">
         <Label fx:id="messageLabel" styleClass="success-message"/>
